refactor: unify Volume metadata management and strengthen config validation
- Master config: add validation for raft_id>0, peers non-empty, ip and
  advertise_addr required; remove hardcoded defaults from main.rs
- Volume struct: remove redundant free_space/next_offset RwLock fields,
  unify statistics from RocksDB allocation CF (single source of truth)
- Deletion strategy: unify to hard delete via delete_needle_atomic;
  store full NeedleInfo in deleted CF to support restore_needle_atomic;
  add purge_expired_deleted for GC cleanup
- write_needle_atomic/compact_cleanup now return AllocationStats to
  avoid extra RocksDB reads
- Startup: sync_allocation_from_index ensures RocksDB allocation CF
  consistency with needle index on crash recovery
- Config files: add ip and net_port to all master configs; fix
  master-single peers from empty to self
- Tests: fix u32→u64 type conversions; update e2e config tests with
  all required fields; update scrub test for hard delete semantics
- Improvement plan: update Phase 3 task statuses

+++ powerfs-cli/src/volume_client.rs
@@ -42,7 +42,7 @@ impl VolumeServerClient {
     #[allow(dead_code)]
     pub async fn create_volume(
         &mut self,
-        volume_id: u32,
+        volume_id: u64,
         size: u64,
     ) -> Result<(), Box<dyn std::error::Error>> {
         let mut service = self.service().await?;
@@ -58,7 +58,7 @@ impl VolumeServerClient {
 
     pub async fn write_needle(
         &mut self,
-        volume_id: u32,
+        volume_id: u64,
         file_key: u64,
         data: &[u8],
     ) -> Result<(), Box<dyn std::error::Error>> {
@@ -85,7 +85,7 @@ impl VolumeServerClient {
 
     pub async fn read_needle(
         &mut self,
-        volume_id: u32,
+        volume_id: u64,
         file_key: u64,
     ) -> Result<Vec<u8>, Box<dyn std::error::Error>> {
         let mut service = self.service().await?;
@@ -106,7 +106,7 @@ impl VolumeServerClient {
     #[allow(dead_code)]
     pub async fn delete_needle(
         &mut self,
-        volume_id: u32,
+        volume_id: u64,
         file_key: u64,
     ) -> Result<(), Box<dyn std::error::Error>> {
         let mut service = self.service().await?;
